Extend package's entries if `EXTEND_PACKAGE` is included in the target's entries
Summary:
PACKAGE files and package() rules have allowed visibility and within_view rules to scale well.

However, in the V1 specification, we decided that in order to provide control of whether attributes (visibility/within_view) are applied or not, if a target defined their own attributes, the package rules' entries would not be applied.

This works for applying rules broadly. But within a package, there are occasions where we want to extend the package's defined entries, not overwrite them. In the [original proposal](https://fb.quip.com/tQu9AngKSFBO), the idea was to have a keyword `__pkg__` (or similar) that referred to the current package, and targets could include that value in it's entries to insinuate, include the package's values and extend with these additional ones. This idea was shelved in order to wait until it was needed.

We're finding that we do now need way to extend the existing list, without overwriting it. To support this functionality, we check whether "EXTEND_PACKAGE" is an entry, and if so, extend the package's entries with those of the target.

Reviewed By: stepancheg

shipit-source-id: b4aa89ce5474a5223b98bdb30f5c3866c13e5009