Provide information in the visibility error message on where rules are defined
Summary:
There has been some feedback that when visibility/within view are violated from rules inherited from PACKAGE files, its not clear where rules are applied. A typical error might be:

```
//foo:Lib6 depends on //baz:Lib1, which is not visible.

More info at:
https://dev.buck.build/concept/visibility.html
```

This was sufficient prior to PACKAGE/package() functionality, as visibility and within_view had to be defined on the target. Going to the BUCK file for the target was sufficient to begin diagnosing the problem. However, with packages, they could be defined several directories up the file path. To make these errors more user friendly, we can update the error message where if the visibility rules are applied from somewhere besides the target's BUCK file, we output the files in which rules are being applied from:

```
//foo:Lib6 depends on //baz:Lib1, which is not visible. Rules defined at:
baz/PACKAGE

More info at:
https://dev.buck.build/concept/visibility.html
```

Reviewed By: stepancheg

shipit-source-id: cd32db9527bc0017ebcad2208d25157c1cc6c1fe