Make definingPath in VisibilityPatterns relative
Summary:
`definingPath` (currently unused) on `VisibilityPattern` is currently typed as a `Path` (and happens to be a mix of absolute and relative, although intended to be relative). Up the diff stack, we make a change to output the defining path in an error message, and instead of an absolute path, it's more user friendly to output a relative path to the cell.

Add typing and make relevant changes to pass the relative path (in many instances, AbsPath was being downtyped to Path).

Reviewed By: stepancheg

shipit-source-id: f7b2e0d3ab9416bdd92d179085460a748c642a8d