Correctly determine flags for apple_binary
Summary:
# Problem

`apple_binary` has broken logic in how it determines compilation flags. Specifically, if `//a:target` is an `apple_binary` with `target_sdk_version` value, then `target_sdk_version` will be ignored.

This happens because a difference in how `apple_binary` determines the platform (i.e., `AppleCxxPlatform`) and how the underlying `CxxBinaryFactory` determines the same platform. The former just looks at the target and if there's no explicit platform (e.g., `iphonesimulator-x86_64`), it will not set minimum deployment target. The latter uses a fallback mechanism.

The above means that the output from `buck build //a:target` is different than `buck build //a:target#platform`, even when `platform` is inferred to be the same in the format command.

# Bug

When preparing to compute compiler flags for `apple_binary`, `AppleDescriptions.populateCxxConstructorArg()` runs which has the following code:

  if (appleCxxPlatform.isPresent()) {
    ... use arg.getTargetSdkVersion() ...
  }

That means that for `apple_binary` without explicit platform flavor in the target name, no deployment target will be set, *even* though the sources will be compilated for the appropriate platform. Sources get compiled correctly because the code uses `CxxPlatforms.getCxxPlatform()` which contains fallback if there's no explicit platform.

# Changes

- Add a config option to control new behavior (`apple.binary_uses_fallback_platform`), off by default.
  - There's no new behavior, by default.
- Add utility method to get `UnresolvedAppleCxxPlatform` using fallback logic
- Use the new method with fallback to determine platform for `apple_binary`

As the previous method was used in a few places, new behavior is behind a config option which will be removed once we've verified the new behavior.

# Config

Reviewed By: blackm00n

shipit-source-id: 60644c2f190b5fdb6b864ed11eacf3d1f29073a5