Remove http_artifact_fetch fallback
Summary:
https://fb.prod.workplace.com/groups/buckeng/permalink/4127394860642431/

Years ago, Buck switched to fetch multiple artifacts at once, resulting in big speedups from a reduction in fetch times. However, looking at the "Buck Build Steps" Scuba table, we still occasionally see the http_artifact_fetch category show up, and the performance in these cases is terrible.

For example, for fb4a first-build-after-rebase, instead of spending ~2 minutes fetching using multi-fetch, we end up spending ~11.5 minutes fetching. Over the whole codebase, the difference is even greater.

I assume that these fallbacks occur when network connectivity is poor. It looks like the fallback was added in D5664760 to assist with the transition to multi-fetch. Is there any situation when it is actually better to be doing single fetches? I assume not, so I'm thinking it would be better to get rid of the fallback?

Reviewed By: cjhopman

shipit-source-id: cc9639b19063e1f1d2c52c4915fb014dda846c4d