Fix robolectric tests running with an external runner on Java 11
Summary: On Java 11 unit tests, we pass in an arg file with the classpath that we want to use for the test (see D15966552, which added this support). Unfortunately, this broke for Robolectric tests with Java 11 and an external test runner, because `RobolectricTestRule` didn't call the `super` method when overriding `onPreTest`. Now that we do that, the arg file contents will actually be filled in, and not empty.

Reviewed By: asp2insp

shipit-source-id: cf5aa0da034d6077265fc8148508c37cf8e35a51