audit actiongraph handle rules being returned multiple times
Summary:
`ActionGraph.getNodes` is documented as sometimes returning nodes more than once. One such case is when
an `alias` exists in the target graph. Some callsites that are sensitive to duplicates explicitly deduplicate;
this change adds `AuditActionGraphCommand.dumpAsJson` to that list. (`dumpAsDot` already does this via
`MutableDirectedGraph.addNode` adding the node to a set.)

Reviewed By: cjhopman

shipit-source-id: 8cdf739a9adbd98b57fcb1ec43a870dce546dd7b