Creating dependencies for detected undeclared deps
Summary: Creating a dependencies for detected module dependencies. In our current state, we force the user to declare a module dependencies in order to for voltron to handle downloading dependencies. a recent example of this is when assistant implicitly depended on assistantclientplatform. in the code, it was assumed that when assistant was downloaded, assistantclientplatform was also safe to use, leading to a CNFE.

Reviewed By: longinoa

shipit-source-id: 07b153cc0ee363eae9dd5c239858738c58b97b7b