Detecting non-transitive seed module to seed module dependencies
Summary:
Adding method to get undeclared non-transitive seed module to seed module dependencies.

The general idea of the algorithm is that we calculate our module dependency graph and match that against our declared dep graph. every edge in the module dep graph should exist in the declared dep graph. this also ensures that we do not ask the user to declare transitive deps.

to follow up handling of the undeclared deps, I posted 2 rfc diffs D22940300 and D22939991.

Reviewed By: longinoa

shipit-source-id: 5d5c96e3f3fd6caa67ba7035bcfe1f2ea4d31af7