Set target-specific deployment version for dylibs (apple_library)
Summary:
# Problem

If `target_sdk_version` is set on an `apple_library` which is used as a dylib, then the sources will be compiled with the `target_sdk_version` but the executable will _not_ be linked with the correct SDK version.

That is, we need to pass `target_sdk_version` both to the compiler and linker.

# Solution

We guard the new behavior behind a flag (`apple.target_sdk_version_linker_flag`) which will be removed once new behavior has been fully deployed.

The changes in this diff add the require SDK flag to the linker if a particular target overrides the SDK version.

Reviewed By: blackm00n

shipit-source-id: 23c4af5df9c4a3c872cb6988d5c0e3b030bc49f3