Use provided target SDK ver for Apple bundles
Summary:
When setting a custom target_sdk_version on an apple_binary, apple_test and apple_library, the SDK version should be propagated to the Info.plist correctly.

Previously, this behavior was broken which now works. This allows the automatic decoupling of target SDK from the global value.

NOTE: If the Info.plist key (`MinimumOSVersion`) is already provided, this code would not override it. It effectively acts as a default value if the key is missing.

Reviewed By: williamtwilson

shipit-source-id: 090f26b74cd6da8dba9dd67a0c721b6484ac61aa