Rename ChromeTraceEvent to ChromeTraceData
Summary: Downward API has a `ChromeTraceEvent`, and buck has a Java representation of a `ChromeTraceEvent`, but they do not map to each other. Buck's `ChromeTraceEvent` is also not a `BuckEvent`, and instead is a data class that holds data to be translated into a separate event concept that chrome traces can understand. Let's rename `ChromeTraceEvent` to `ChromeTraceData` since this class uses "event" not in the buck sense, and it's easy to confuse it with downward API's `ChromeTraceEvent`

Reviewed By: bobyangyf

shipit-source-id: 5b3dec695e748c68f4d703dbd460b5be4edacf54