Restore caching of kotlinc
Summary: This was turned off in D20655285 because we were seeing issues with stale jars being used. That bug was fixed in Java 9 so now that we are using the Java 11 compiler we should be able to turn this back on. If we get any reports of problems we'll just have to turn it off again.

Reviewed By: asp2insp

shipit-source-id: 385dc29aecfcc06228aa7cc2fe9f4cf70e38ab6b