Close ClassLoaderCache in JarBackedReflectedKotlinc
Summary: I'm investigating the bug we were seeing where Kotlinc was incorrectly keeping a jar in its cache when it should have updated it to match the one on disk (https://fb.prod.workplace.com/groups/392546017460686/permalink/3462653623783228/). I noticed that when we get the ClassLoaderCache we add a reference to it, but we never remove that reference. I can't really see how this would contribute to the bug (which is currently mitigated by turning off caching), but I guess it is worth fixing anyway?

Reviewed By: asp2insp

shipit-source-id: 84233f476bfc579a214924a1606fd25648857ddd