Escape/quote cmd.exe special characters in genrule macros and environment variable substitutions
Summary: Leveraging new cmd.exe ecaping/quoting functionality introduced in a previous change to escape/quote special characters in genrule macros and environment variable substitutions. Allows, for example, the `$(exe)` macro to supply an executable path that contains `','` (used by Buck as a flavor separator, but also is a cmd.exe special character that causes cmd.exe to not see the entire path) without requiring explicit quoting in the genrule.

Reviewed By: akrieger

shipit-source-id: c315401d2b805ab6c15f1a299f4605c0ef6d3da5