Create option to turn js_library into a dummy rule when missing a transform profile flavor
Summary:
Step 3 of the plan outlined in the summary of D22316860 - see that diff for more details.

This diff introduces a temporary Buck config flag (`js.require_transform_profile_flavor_for_files`) that will be removed in an upcoming diff (D22397250) once we prove that setting it to true is safe for all workflows.

When this flag is set:

1. Unflavored `js_library` rules cannot be built (same as D22397251, different enforcement mechanism).
2. Unflavored `js_library` rules will not generate `js_file` dependencies.

Reviewed By: christolliday

shipit-source-id: ea5b18d3bcfc07f94963b57022aa83bd744509b3