Add macro resolving to public header tree
Summary:
Support a preprocessor flag macro for `cxx_library` rules which
expands to the location of the public header tree (if it exists).

In modular builds, we build the module file in a temp directory
but use it from via a module map file installed in the symlink tree
of the C++ library rule that "owns" it.  Normally, clang detects
the correct "module home" dir by finding these module map
files in the header search path.  However, this breaks when used
with header maps.

This diff will instead allow explicitly loading module map files
so that we eagerly discover the correct "module home" paths
to use for modules and allows us to avoid current hacky work
arounds and work in
cross-cell builds.

Reviewed By: mzlee

shipit-source-id: 8b8111d693850256bf90ef6ef95c6492801ab2a4