Filter current node when generating SWIFT_INCLUDE_PATHS
Summary:
Currently, we add the current node's paths to the `SWIFT_INCLUDE_PATHS`. So, library X will have the include path for library X included. This seems wrong, since the current module can't import itself. So, filter it out. I don't know if it's //actively// breaking things, just seems wrong.

It might be necessary to add the current module's path for mixed libraries. However, this would be the wrong one, we would want the "underlying module" path in that case (`#apple-swift-underlying-module`), not the "public" flavor.

Reviewed By: milend

shipit-source-id: 754d422cad6a15186c33d6b4e3d65f0a4be82442