Pass correct platform name to actool for Mac Catalyst
Summary:
Up until now, the Buck platform names coincided perfectly with the platform names passed to the Xcode toolchains. But with Mac Catalyst, that's no longer the case which has an internal name of `maccatalyst` and external name of `macosx`.

In this diff, we add an `getExternalName` field to `ApplePlatform` which is used to pass the correct platform to `actool`.

(Note: this ignores all push blocking failures!)

Reviewed By: blackm00n

shipit-source-id: c70625a4737693c8afc1b0f279a9de65cd4b4f5b