Add support for resources device family
Summary:
Add an optional device family to `AppleSdk` which will be used to pass the correct `--target-device` to `actool`. The target device value is a property of the SDK being used, so it's much more flexible and correct to be parsing that info from the SDK plist rather than relying on hardcoded values in Buck.

This will be used to correctly compile resources for Mac Catalyst apps.

(Note: this ignores all push blocking failures!)

Reviewed By: blackm00n

shipit-source-id: efa7bcc82ddad07424eabd69c38a9693e0fb2c47