Make SourcePaths' rulekeys aware of output labels
Summary:
Consider a rule `foo` with output labels `bar` and `baz`, each pointing to a different output

When used in `srcs`, `foo[bar]` and `foo[baz]` currently have the same rulekey. This is because the rulekey factory currently does not take the output label into account when handling build target source paths

This change makes the rulekey factory know about the output label when handling source paths by relying on source paths' `representationForRuleKey`. Initial approach was to make handling build target source paths more similar to path source paths (i.e. set the path directly onto the hasher), but this does not handle the case where two different build targets have the same underlying path. Another option considered was to add  `putBuildTargetSourcePath` to `RuleKeyHasher`. This change is more involved, and would require changing several `RuleKeyHasher` representations, as well as some generated code

Note that `buck build :foo[bar]` and `buck build :foo[baz]` will still result in the same rulekeys, because `:foo[bar]` and `:foo[baz]` by themselves are not source paths. This is not really correct, but it is practical. Currently we build all of `foo` anyway even if the user only requests `bar` output, so it's better from a caching perspective for `bar` and `baz` to point to the same things and grab each others' outputs

Reviewed By: bobyangyf

shipit-source-id: 9a2167f95b2cb7f7e5669d40ba93d203bfc823e6