update run_inplace.py to include the program name near the start
Summary:
When running an in-place style python binary, the process arguments are
`python -c STARTUP_CODE ACTUAL_ARGUMENTS`.

This `STARTUP_CODE` block is a very large and unwieldy block of Python code.
As a result, when viewing the running processes in `top` or on `ps` the main
contents that get seen are
`python -c "# Wrap everything in a private function to prevent..."`
This makes it pretty awkward to actually tell what is being run.

This diff attempts to make it slightly easier to tell what is being run by at
least including the python script's name at the start of this comment block.

Reviewed By: thatch

shipit-source-id: 9b5249369ac48d2257f1c694b35fe655e42e4547