apple | Introduce rule to compute non-processed resource hashes
Summary: Processing happens in separate rule - `AppleProcessResources`, where hashes for processed resources should be computed. In case of resources which should not be processed we have to compute their hashes in auxiliary rule which I'm introducing in this diff. Hashes are written in JSON form: "path relative to bundle root" -> "hash of file or directory". This JSON file will be used later to understand if resource is changed and newer version should be moved to result bundle.

Reviewed By: milend

fbshipit-source-id: b0bb084ef4b767a54f16596c4421217ea89525ac