apple | extract AppleBundleComponentCopySpec into public class
Summary:
Also add `ignoreIfMissing` property to it to handle when source file/directory is missing for some reason (used during dry signing) and `createCopyStep` method to create proper build step as per spec.

Naming `*Component*` is intentional to not confuse with `AppleBundlePart` classes, the introduced class is an adapter to `SourcePathWithDestination` as well.

Reviewed By: milend

fbshipit-source-id: d43591c46fc81d3ce9f6566a306b38d4df7c6128