AppleDebuggableBinary: hoist flavor addition up one level
Summary:
The `AppleDebuggableBinary.create*` functions were appending flavors to a provided `BuildTarget` before passing it in to the constructor. That's an anti-pattern -- whatever `BuildTarget` is passed to `ActionGraphBuilder.requireRule` should be the one that's used to construct a rule, otherwise the `ActionGraphBuilder.getRule` won't be able to find it.

This commit takes a baby-step toward fixing the issue by hoisting the code to append the flavors up to the single callsite of these methods. This does not change observed behavior; that'll happen in a later commit.

Reviewed By: milend

fbshipit-source-id: 20b25ffb1b960e2bf6c60c304b3dda06f99ca038