Fix AppleDebuggableBinary target names
Summary:
`AppleDebuggableBinary` was adding flavors to the target name provided, which often resulted in it using a different target name than was actually requested. For example, in one case `ActionGraphBuilder.requireRule` was called with `foo#dwarf,iphonesimulator-x86_64,strip-non-global`, but the `AppleDebuggableBinary` was created with `foo#apple-debuggable-binary,dwarf,iphonesimulator-x86_64` as its build target. The result of this is that calling `getBuildTarget` or `getSourcePathToOutput` (which in turn calls `getBuildTarget`) on the rule will return a target name that, as far as the action graph is concerned, does not exist.

This manifested as crashes in `audit actiongraph`, which attempts to resolve the output source path to an absolute path on the filesystem and fails because there is no entry in the action graph's rule map that matches what the `AppleDebuggableBinary` reports as its target name. However, this sort of thing (creating a rule whose own idea of its name doesn't match the action graph's idea of its name) can cause other problems as well.

This change adds a test that demonstrates the issue, and fixes it by adding a call to `ActionGraphBuilder.computeIfAbsent` around the code that actually creates the rule after adding flavors to the provided `BuildTarget`. This ensures that the actiongraph can look up the rule by the name it knows for itself (in addition to by whatever name was actually passed to `requireRule`).

Reviewed By: milend

fbshipit-source-id: ae9cfe46b2c751f6bf9416ae65ee510e38a87764