Automatically include transitive deps in jni_deps in java_test
Summary:
Previously, we would only populate the symlink tree with explicitly
whitelisted rules (if a whitelist was present).  However, this requires users
to list all transitive dependencies explicitly, which is cumbersome and
sometimes requires expanding rule visibility.  Now, we automatically add all
transitive declared dependencies.

I don't think a user would ever care that a "normal" dependency was added this
way, since they are most likely needed in order for the test to run.  It's
possible that we might traverse a "code-generator-like" dependency, but since
`requiredCxxTargets` is only used as a filter for targets already in the
initial symlink tree, I don't expect any runtime issues.

Reviewed By: cjhopman

fbshipit-source-id: b41cc1015c26a75a20b0cfda537cbe5f84f74057