Framework path function returns optional
Summary:
# Summary

There's a function which transforms a `FrameworkPath` (i.e., what appears in the `frameworks` field of `apple_library` and `apple_binary`) into a framework search path (i.e., passed as `-F` to the linker).

At present, the function *always* returns a `Path` but that's problematic when we need to compile Catalyst apps because the current logic ends up including a search path to the wrong architecture

# Changes

In this diff, we change the function so that it *can* return an `Optional<Path>`.

NOTE: There's no changes in behaviour in this diff, only adjustment of the type signature.

Reviewed By: blackm00n

fbshipit-source-id: 35dcb55db27d711c27cbdfbc835fd79e0bd43f15