Add option to skip system framework search paths
Summary:
# Buck: Framework Search Paths

When frameworks appear in the `frameworks` field for `apple_library`, `apple_binary` and `apple_test`, Buck does two things:

- Add the *parent* dir of the framework to the framework search paths.
- Add the framework name to the parameters

For example, if we had the following:

  apple_library(
    ...
    frameworks = ["$SDKROOT/System/Library/Frameworks/MapKit.framework"],
    ...
  )

Then two parameters would be passed to the linker/compiler:

  -F /Applications/Xcode/...Name.sdk/System/Library/Frameworks -framework MapKit

# How Linker Finds Frameworks

The linker has a set of built-in search paths which are used to locate frameworks.

> ld maintains a list of directories to search for a library or framework
>  to use.  The default library search path is /usr/lib then /usr/local/lib.
>  The -L option will add a new library search path.  The default framework
>  search path is /Library/Frameworks then /System/Library/Frameworks.

> The -F option will a new framework
> search path.  The -Z option will remove the standard search paths.  The
> -syslibroot option will prepend a prefix to all search paths.

Note that we *always* pass `-syslibroot` (as `-isysroot` via the Clang driver), so that we can compile for different platforms (iOS/macOS/watchOS/tvOS).

### Corollary

The above means that, for any system frameworks, we do not *need* to be passing the `-F` flag with the parent dir. In fact, it actually leads to broken behaviour for Catalyst apps.

# Catalyst & Broken Behaviour

In order to link and compile Catalyst apps, the linker needs to find the *Catalyst* versions of the system frameworks. But note that, to compile Catalyst apps, we still need to use the macOS SDK (i.e., `-isysroot` will point to the macOS SDK).

### How Linking Should Work

In both Xcode and Buck, when building Catalyst apps, we pass two relevant flags to the Clang driver:

  -isysroot /.../Developer/SDKs/MacOSX10.15.sdk -iframework /.../MacOSX10.15.sdk/System/iOSSupport/System/Library/Frameworks

Those driver flags become `-isysroot` -> `-syslibroot` and `-iframework` -> `-F`. Note that all the Catalyst frameworks reside in the `System/iOSSupport` folders. So, in effect, the above two parameters mean that any system framework references would firstly be resolved against the Catalyst versions.

### What Happens Instead

The problem arises because Buck will add the parent dirs of the system frameworks to the framework search paths. It would insert `-F /.../Developer/SDKs/MacOSX10.15.sdk/System/Library/Frameworks` but that would be *wrong* because then the linker will find the macOS flavor of the framework, not the Catalyst one (as later `-F` arguments take precedence over earlier ones, i.e., the `System/iOSSupport` one).

# The Changes

The change is quite simple: we should *not* be adding search paths for system frameworks, that would mean they get resolved correctly to the appropriate one.

In this diff, we add a parameter to the functions that generate those parameters whether to skip adding search paths for system frameworks.

NOTE: In this diff, there are no behavioural changes, all pre-existing behaviour is preserved.

Reviewed By: blackm00n

fbshipit-source-id: 4f26dbf63602143cf667624d2ec2e1cf2f6c0e9d