Use modulemaps for vfs symlinks only when isModular
Summary:
A slight oversight in my previous implementation for VFS overlays is that we were presuming all dependent overlays as being modular.

This assumption forced *every* vfs module to create a modulemap even if it wasn't required. This led to `buck build` breaking for mixed module/non-module frameworks to break when `--config apple.swift_uses_vfs_overlays=true` was appended to the `build` command.

In this diff, I check if the current target `isModular()` before instructing it to generate its output.

Reviewed By: natestedman

fbshipit-source-id: e8dea02195c366e2059369b7359ffb2d1b2c2ad5