buck/rust: don't include proc_macros in the transitive deps
Summary:
Procedural macros (proc-macros) are essentially compiler plugins. They're
treated as normal crate dependencies in many ways, but they are only required
for compilation, not a final link stage. In particular, they are never
candidates for native linking.

This diff cleans things up so they are not propagated too far in the linker
dependencies.

Reviewed By: cjhopman

fbshipit-source-id: cce3a5b87299473365f45d6aae4801a83da41bab