Propagate Swift target triple correctly for Catalyst platform
Summary:
The Swift target triple which was being passed for Catalyst apps was not computed correctly. This diff fixes that issue, so Swift code can now be used in Catalyst apps.

# Behavioural Change

The values of the two fields now change when compiling Swift code:

- Platform:
  - Previously: `appleSdk.getApplePlatform().getSwiftName().orElse(appleSdk.getApplePlatform().getName())`
  - Currently: `AppleSdk.getTargetTriplePlatformName()` if present, otherwise previous value before this diff.
- ABI
  - Previously: `Optional.empty()`
  - Currently: `AppleSdk.getTargetTripleABI()`

NOTE: For *all* SDKs except the Catalyst one, all the currently used values are *exactly the same* as the previously used one. That's because for those SDKs, both `AppleSdk.getTargetTriplePlatformName()` and `AppleSdk.getTargetTripleABI()` return `Optional.empty`. Thus, only the Catalyst Swift compilation command will see a change in behaviour.

Reviewed By: blackm00n

fbshipit-source-id: 9ca76f7435e422927d111130449bebb6007ebcae