Add iOS twin prefix Swift runtime paths for Catalyst apps
Summary: When linking against the Swift runtime for Catalyst apps, we need to pass two different library search paths: the Catalyst specific one and the primary one that contains most of the Swift runtime (e.g., `libswiftDarwin`). The order is important as for some libraries/overlays, there are both Catalyst and macOS ones, so the Catalyst search path must come first.

Reviewed By: blackm00n

fbshipit-source-id: 7c86b1bc42d0b6f7af6b7c91d0d2cff7af6cc12e