When parsing for `show-target-hash`, provide cell
Summary: When doing `buck targets --show-target-hash` on a cross-cell node, the select parsing does not resolve the correct cell. In `ParserWithConfigurableAttributes::resolveConfigurableAttributes`, we try to create a selector list with a relative path that does not match the `CellNameResolver`. Thus, the resolver fails to find the correct path to the selector file.

Reviewed By: stepancheg

fbshipit-source-id: 5f4a580eaab57e250fa0b3f5a3d684ac05fcbed1