buck/rust: separate plugin platform config
Summary:
Add rustc_plugin_platform option for configuration. This specifies a
separate platform to use when compiling compiler plugins (eg, procedural
macros). These must be compiled with options which are compatible with how
rustc itself has been compiled - same architecture, similar optimization
level, sanitizer settings, etc.

If this is not specified then it continues to use a single plaform for
both the target and plugins.

Reviewed By: cjhopman

fbshipit-source-id: a4430a5f8a3287f6f3913b128297191432cc5782