Cache Mach-O export tries
Summary:
If a dylib gets changed, then *all* dependents need to compute the dylib's export trie. This means that every single dependent will perform the same expensive computation.

In this diff, we introduce a cache, so that we do not compute keep recomputing the same export trie multiple times.

(Note: this ignores all push blocking failures!)

Reviewed By: blackm00n

fbshipit-source-id: e971d55ee08363a25d1c2d276ccf8734abef1e34