buck: fix jitpack on dev branch
Summary:
The jitpack build on the dev branch was broken. I just copied
jitpack.yml from master and it worked.

This is blocking a working release of `facebookincubator/antlir`, which uses
buck to build the open source project and requires some buck features that are
not on `master` (notably `native.sha256`).

Reviewed By: dtolnay

fbshipit-source-id: dbc4c13dd12073eccd5267af8f7d769588567767