Move vfsoverlay content-external to use relative path
Summary: This moves VFSOverlay to use a path relative to the filesystem rather than an absolute path. The motivation for me is that on my system, a bad absolute path was being used. A more broad motivation is that absolute paths should be avoided where possible since they break caching and RE.

Reviewed By: scottrice

fbshipit-source-id: 9a1c82d8d5aa830c7e3aae101c9714aec0949a1b