Add BuildableCommandExecutionStep
Summary:
Migrated buildables that implement `BuildableWithExternalAction` will return a `BuildableCommand` according to the interface. That `BuildableCommand` will then get wrapped into this `BuildableCommandExecutionStep` and get passed to the external binary. This step is essentially a bridge between the current execution flow and the external binary

Side note: constants are split out from the other parts of `external/utils` to avoid a circular dependency

Reviewed By: bobyangyf

fbshipit-source-id: ed5586e0fcb766b774993e985f78d89240ad7e5e