Update packages for compatibility with macOS 11 Big Sur
Summary:
# Context

macOS 11 adopts the [dyld shared cache](https://iphonedevwiki.net/index.php/Dyld_shared_cache) which means that the system libraries do not actually exist on the filesystem. This broke JNA and it was fixed in late June 2020 ([issue#1215](https://github.com/java-native-access/jna/issues/1215)).

# Changes

Because we have other third-party JARs which depend on JNA, we had to update those. In summary, we updated all of the following modules:

- JNA
- nailgun
- NuProcess
- oshi
- Java Obj-C Bridge

Reviewed By: blackm00n

fbshipit-source-id: f29c775912ac3a84ec5615f6999b7bccf212ead7