Loosen some visibility restrictions
Summary:
These things are specifying specific targets in `//programs:` that can
depend on them, but that's just encoding internal details of how the buck
package is built into these visibility specs. It should instead allow anything
in the package to depend on it so that that package is free to change its
structure as it wants.

Reviewed By: stepancheg

fbshipit-source-id: 53a86b35e74ff31419240560b1dae4ec54594c32