Add a simple files signature helper
Summary:
This is very similar to the previously used hash_files (in py/hash).
It changes it to accept a name->file mapping and to support directories.  Also,
produces a richer signature, including both the top-level signature of
everything and a breakdown of all the individual signatures.

We'll use this for resources in the buck pex to determine when they need to be
refreshed.

Reviewed By: stepancheg

fbshipit-source-id: c61a51ae2da8c3da830897f06709bfef1b4c542d