Compute and embed a signature of all resources in the pex
Summary:
This computes a signature over the entire set of resources. This will
be used (instead of the source control version) to determine reuse of resources
or the daemon. This will more correctly handle the local development case.

This signature will not be used for rulekey calculations. We will continue to
use the version for that.

Reviewed By: stepancheg

fbshipit-source-id: 9c111afbcf9256735e8c57999db9037d347a6509