Delete buck-local
Summary:
This is intended to help deal with the fact that buck reads the hg
commit version and uses that to determine if it can skip work. That's a problem
for local development where the version doesn't change.

buck-local is an inefficient approach to dealing with that problem. I'll
rewrite this to handle it better, but it requires some refactoring that is
easier if buck-local is deleted first.

Updated the alias to point to buck to not break user's workflows (and with the
rest of this stack it'll have the right behavior that they expect from
buck-local).

Reviewed By: Skory

fbshipit-source-id: 28f863ef3d9cabb3445164eb0cb5979eb72705a4