Change resource unpacking to use the resource signature
Summary:
Instead of a bunch of hacks to try to not reuse the wrong resources,
use the resource signature to identify them instead.

The resource signature is a hash of all the resources in the pex and is
included as part of the package_info. That package_info is read by the
python wrapper and then we use the resources signature from that
(instead of the previous buck/fakebuck version) to determine if resources
can be reused.

Since the resources root is also passed as part of the jvm args, this also
fixes the issue with BUCK_FAKE_VERSION reusing the wrong buckd.

Reviewed By: stepancheg

fbshipit-source-id: c0eb620e76aadc50edabf90e8f2a39951df0d74e