Binary output format for target-info.json
Summary:
Adds an option to write binary representations of target info to disk alongside the json one.

This binary format is optimized for fast, on-disk lookups in a hashtable-like datastructure within a random access file.

The format contains a header section and a body section. The header section is an ordered table mapping hashcodes of the key to offsets within the body section. The body section consists of records of the value data. Since each hashcode can map to > 1 value, there may be multiple values at a given record. Lookups involve first a binary search through the table in the header, followed by a linear scan of records matching the hashcode.

We also write out a file that maps module names -> buck targets for reverse lookups.

Reviewed By: cute-jumper

fbshipit-source-id: aa42cbbec12c8adf642d28a03f37386ae0f62030