Add buck_build_id to BuckRepo and BuckResult
Summary:
We can set BUCK_BUILD_ID in the environment before running a buck command to get buck logs with the same id.
This diff makes `BuckRepo` generate a uuid and set it as BUCK_BUILD_ID before running the command.
The resulting `BuckResult` will also carry the same id.

Reviewed By: bobyangyf

fbshipit-source-id: ae64e0ffaacabf54170d432a9ba98d81480bce32