Ignore unknown fields from idb for AppleDevice
Summary:
The JSON output from `idb` seems to depend on various factors and sometimes it includes a `device` field which Buck does not know about (nor need at present, it's just device metadata). This breaks the ability to use `buck install` with Xcode 12 and a physical device (iPhone XS).

In this diff, `AppleDevice` deserialization is modified, so that unknown fields can be ignored.

Overall, the current coupling between Buck and idb makes it much harder to evolve the `idb` output due to the hard dependency on exact format.

Reviewed By: blackm00n

fbshipit-source-id: b48e210803595f218ea8c81ed75ced5c1a941ac5