Resolve conflict between modular libraries and merged header maps
Summary:
To populate Xcode's `HEADER_SEARCH_PATHS` argument for indexing and for building tests, we create a merged header map of all dependencies in the top-level target. This is used for all libraries, so libraries will have headers that are not from their dependencies available in this context. However, this creates an issue for modular imports: headers end up being "duplicated", because the entry in the header map is treated as separate from the header in the module. The current effect of this is that indexing breaks for Swift code for which this duplicate issue appears.

To resolve this, we need to remove these entries from the header maps. To reduce the risk of breaking indexing, I've chosen to do this only for libraries that are themselves modular, so we now create two header maps: the existing one for all libraries, and a new one for all non-modular libraries. Modular libraries receive the new header map, while non-modular libraries are unchanged.

Next, we need to provide the headers in another way. We were already adding the modular symlink tree paths to `SWIFT_INCLUDE_PATHS`, so I've moved it to `HEADER_SEARCH_PATHS`, and left only `.swiftmodule` paths in `SWIFT_INCLUDE_PATHS`. Xcode will automatically pass the contents of `HEADER_SEARCH_PATHS` to the Swift compiler as well, and now it will also pass it for Objective-C code.

Reviewed By: milend

fbshipit-source-id: 6da3278964c89e0eabd3951b91232ac1333c55d6