fix VFSOverlay for ObjC headers declared with a map
Summary:
When I first wrote the VFS, i was unaware of the way `exported_headers` would generate a nested file structure inside `buck-out/` if the input used a map instead of an array (flat structure).

This assumption of an always flat file structure created a mismatch between with the modulemap codegen and what was reflected in the file structure which would be consumed by the VFS overlay I was creating, thus breaking the reverse dep.

I also added a new test file structure to catch this.

Reviewed By: williamtwilson

fbshipit-source-id: f9bfea2345330eff5b70f4a769a110b728b6f82d